Edge - rename EdgeView
