+++ apps/AEPSampleApp/App.js
@@ -7,7 +7,7 @@ import Core from './extensions/Core';
 import Identity from './extensions/Identity';
 import Messaging from './extensions/Messaging';
 import EdgeIdentity from './extensions/EdgeIdentity';
-import EdgeView from './extensions/Edge';
+import EdgeView from './extensions/EdgeView';
 import Assurance from './extensions/Assurance';
 
 function HomeScreen({ navigation }) {
