+++ apps/AEPSampleApp/extensions/EdgeView.js
